Remove all neutron agents from controller metal
With only a single flat interface, neutron-linuxbridge-agent breaks
the controller networking, taking all controllers offline.

neutron-linuxbridge-agent unbinds bond0 from the br-mgmt bridge and
attaches it to its own bridges instead, which breaks the host
networking.

This is not an issue when the flat network is passed to the agents
container, since the container tap is attached to the br-mgmt
bridge on the host, and neutron can hang its bridges off the tap
interface in the container.