Use utility container for service bootstrapping
Configure OSA to use the utility container for bootstrapping OpenStack
services. In Rocky, OSA began using the deployment host by default
instead of the utility container. This breaks our deployment model
because the deployment host does not have access to the internal
OpenStack API endpoints.

Revert back to the previous behavior of using the utility container
for bootstrapping services.

Change-Id: Iebfb6583c1b02bdc7422fb7c3fbdaf3a851aec43